symbols/pyproject.toml
[project]
name = "symbols"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.1.0",
    "click>=8.1.8",
    "flake8>=7.2.0",
    "ipdb>=0.13.13",
    "isort>=6.0.1",
    "pandas>=2.2.3",
    "pillow>=12.0.0",
    "plotly>=6.0.1",
    "pylint>=3.3.6",
    "pyright>=1.1.398",
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
    "ruff>=0.11.7",
    "squarify>=0.4.4",
]
